update the benchmark tests with new class and function
