+++ include/tridiagonal_matrix/trid_matrix.hpp
@@ -3,6 +3,9 @@
 
 #pragma once
 
+#include <iostream>
+#include <iomanip>
+#include <algorithm>
 #include <vector>
 
 template<typename T>
