+++ test/benchmark/CMakeLists.txt
@@ -6,4 +6,4 @@ include(test_util)
 ################################
 # Tests
 
-create_standard_test(NAME performance_naive_trid_matrix_solver SOURCES test_naive_trid_matrix_solver.cpp LIBRARIES linear_algebra benchmark::benchmark)
+create_standard_test(NAME performance_trid_matrix_solver SOURCES test_trid_matrix_solver.cpp LIBRARIES linear_algebra benchmark::benchmark)
